requirements.txt
tensorflow==1.14.0
google-cloud-texttospeech==1.0.1
moviepy==1.0.2
grammarbot==0.2.0
numpy==1.18.1